fix: correct the formula

+++ src/tile/Util.h
@@ -16,7 +16,7 @@ struct BoundingBox {
 float deg2Rad(float degree);
 float rad2Deg(float radius);
 std::tuple<int, int, int> lonLat2Cords(float latitude, float longitude, int zoom);
-std::tuple<float, float> cords2LonLat(int x, int y, int zoom);
+std::tuple<float, float> cords2LonLat(float x, float y, int zoom);
 int bestZoomLevel(const BoundingBox& bbox, int padding, int mapWidth, int mapHeight);
 
 }
